+++ omweather/src/weather-utils.h
@@ -35,7 +35,7 @@ float mb2inch(float pressure);
 void set_font_size(GtkWidget *widget, char font_size);
 void set_background_color(GtkWidget *widget, GdkColor *bgc);
 void set_font_color(GtkWidget *widget, guint16 red, guint16 green, guint16 blue);
-int c2f(int temp);
+float c2f(float temp);
 void swap_temperature(int *hi, int *low);
 GtkWidget* lookup_widget(GtkWidget* widget, const gchar* widget_name);
 #define GLADE_HOOKUP_OBJECT(component,widget,name) \
